requirements.txt
aniso8601==7.0.0
azure-cognitiveservices-vision-computervision==0.5.0
azure-common==1.1.23
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-ckeditor==5.7.1
django-cors-headers==3.1.1
django-easy-select2==1.5.6
django-graphql-jwt==0.2.2
django-js-asset==1.2.2
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
grpcio==1.24.1
idna==2.8
isodate==0.6.0
msrest==0.6.10
oauthlib==3.1.0
promise==2.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
